Refactor the classes and move the templates under template, REST services to service, etc. #20
