--- servicebox-app/src/main/java/io/halkyon/GreetingResource.java
@@ -1,16 +1,0 @@
-package io.halkyon;
-
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
-
-@Path("/hello")
-public class GreetingResource {
-
-    @GET
-    @Produces(MediaType.TEXT_PLAIN)
-    public String hello() {
-        return "Hello from Primaza Service Box App";
-    }
-}
